requirements.txt
python-telegram-bot==22.2
gspread
google-auth
APScheduler
python-dotenv
